+++ Frontend/src/components/toneJS/scales/isomorphicLayout.js
@@ -1,0 +1,12 @@
+
+
+export class IsomorphicLayout{
+
+    constructor() {
+    }
+
+    coordinateToNote(valueX, valueY){
+        //console.log('ISO:' + valueX, valueY)
+        return "C4"
+    }
+}
